Improved status display list of attached archive files in openHistorian output adapter. General global code cleanup.

+++ Source/Libraries/GSF.SortedTreeStore/IO/FileStructure/IndexMapper.cs
@@ -5,10 +5,10 @@
 //
 //  Licensed to the Grid Protection Alliance (GPA) under one or more contributor license agreements. See
 //  the NOTICE file distributed with this work for additional information regarding copyright ownership.
-//  The GPA licenses this file to you under the Eclipse Public License -v 1.0 (the "License"); you may
+//  The GPA licenses this file to you under the MIT License (MIT), the "License"; you may
 //  not use this file except in compliance with the License. You may obtain a copy of the License at:
 //
-//      http://www.opensource.org/licenses/eclipse-1.0.php
+//      http://opensource.org/licenses/MIT
 //
 //  Unless agreed to in writing, the subject software distributed under the License is distributed on an
 //  "AS-IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. Refer to the
@@ -57,9 +57,9 @@ namespace GSF.IO.FileStructure
 
             m_addressPerBlock = (uint)(blockSize - FileStructureConstants.BlockFooterLength) >> 2;
             m_addressPerBlock = (uint)Math.Min(uint.MaxValue, (ulong)m_addressPerBlock);
-            m_addressPerBlock2 = (uint)Math.Min(uint.MaxValue, (ulong)m_addressPerBlock * (ulong)m_addressPerBlock);
-            m_addressPerBlock3 = (uint)Math.Min(uint.MaxValue, (ulong)m_addressPerBlock * (ulong)m_addressPerBlock * (ulong)m_addressPerBlock);
-            m_lastAddressableBlockIndex = (uint)Math.Min(uint.MaxValue - 1, (ulong)m_addressPerBlock * (ulong)m_addressPerBlock * (ulong)m_addressPerBlock * (ulong)m_addressPerBlock - 1);
+            m_addressPerBlock2 = (uint)Math.Min(uint.MaxValue, m_addressPerBlock * (ulong)m_addressPerBlock);
+            m_addressPerBlock3 = (uint)Math.Min(uint.MaxValue, m_addressPerBlock * (ulong)m_addressPerBlock * m_addressPerBlock);
+            m_lastAddressableBlockIndex = (uint)Math.Min(uint.MaxValue - 1, m_addressPerBlock * (ulong)m_addressPerBlock * m_addressPerBlock * m_addressPerBlock - 1);
             //initializes all of the values
             MapPosition(0);
         }
