+++ Source/Libraries/GSF.SortedTreeStore/StepTimer.cs
@@ -5,10 +5,10 @@
 //
 //  Licensed to the Grid Protection Alliance (GPA) under one or more contributor license agreements. See
 //  the NOTICE file distributed with this work for additional information regarding copyright ownership.
-//  The GPA licenses this file to you under the Eclipse Public License -v 1.0 (the "License"); you may
+//  The GPA licenses this file to you under the MIT License (MIT), the "License"; you may
 //  not use this file except in compliance with the License. You may obtain a copy of the License at:
 //
-//      http://www.opensource.org/licenses/eclipse-1.0.php
+//      http://opensource.org/licenses/MIT
 //
 //  Unless agreed to in writing, the subject software distributed under the License is distributed on an
 //  "AS-IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. Refer to the
@@ -138,7 +138,7 @@ namespace GSF
             }
 
             list.Sort();
-            return (list[list.Count >> 2] * 1000000.0 / (double)innerLoopCount / (double)internalLoopCount).ToString("0.0");
+            return (list[list.Count >> 2] * 1000000.0 / innerLoopCount / internalLoopCount).ToString("0.0");
         }
 
         private static double TimeLoop(Stopwatch sw, Action del, int loopCount)
@@ -183,7 +183,7 @@ namespace GSF
             }
 
             list.Sort();
-            return (list[list.Count >> 2] * 1000000.0 / (double)innerLoopCount / (double)internalLoopCount).ToString("0.0");
+            return (list[list.Count >> 2] * 1000000.0 / innerLoopCount / internalLoopCount).ToString("0.0");
         }
 
         private static double TimeLoop(Stopwatch sw, Action<Stopwatch> del, int loopCount)
